more memguard debugging
git-svn-id: svn+ssh://frodo/cluster/svn/projects@13540 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/include/psc_util/alloc.h
@@ -154,7 +154,7 @@ void	 _psc_munlock(void *, size_t);
 
 void	  psc_memallocs_init(void);
 
-extern long			psc_pagesize;
+extern int			psc_pagesize;
 
 #ifdef DEBUG
 extern struct psc_hashtbl	psc_memallocs;
